+++ test/document.test.js
@@ -959,6 +959,25 @@ describe('document', function() {
 
       assert.ok(foundGroup.toJSON()._users[0].hello);
     });
+
+    it('jsonifying with undefined path (gh-11922)', async function() {
+      const userSchema = new Schema({
+        name: String,
+        friends: [{
+          type: String,
+          transform(friendName) {
+            return `Hi, ${friendName}`;
+          }
+        }]
+      });
+      const User = db.model('User', userSchema);
+      const alice = await User.create({ name: 'Alic', friends: ['Bob', 'Jack'] });
+      const foundAlice = await User.findById(alice._id, { name: true });
+      assert.equal(foundAlice.friends, undefined);
+      const foundAlicJson = foundAlice.toJSON();
+      assert.equal(foundAlicJson.friends, undefined);
+      assert.equal(foundAlicJson.name, 'Alic');
+    });
   });
 
   describe('inspect', function() {
@@ -8346,7 +8365,6 @@ describe('document', function() {
     });
     const Organization = db.model('Test', organizationSchema);
 
-
     const org = new Organization();
     org.set('name', 'MongoDB');
 
@@ -11453,5 +11471,14 @@ describe('document', function() {
     const baz2 = await Baz.create({});
     baz2.bar = bar2;
     assert.ok(baz.populated('bar'));
+  });
+});
+
+describe('Check if instance function that is supplied in schema option is availabe', function() {
+  it('should give an instance function back rather than undefined', function ModelJS() {
+    const testSchema = new mongoose.Schema({}, { methods: { instanceFn() { return 'Returned from DocumentInstanceFn'; } } });
+    const TestModel = mongoose.model('TestModel', testSchema);
+    const TestDocument = new TestModel({});
+    assert.equal(TestDocument.instanceFn(), 'Returned from DocumentInstanceFn');
   });
 });